update the test app to acknowledge the new 'view' property + renamed folder to test-example-app
