+++ test-example-app/containers/navbar.js
@@ -25,24 +25,26 @@ export default SepCon.createComponent({
                 }
             }
         },
-        render() {
-            let buttons = [];
-            for (let key in this.props.pages) {
-                const page = this.props.pages[key];
-                buttons.push(Button.createTag()
-                    .props({
-                        label: page.title,
-                        isActive: this.props.currentPage.id === key
-                    })
-                    .methods({
-                        onclick: this.methods.navigate.bind(this, page.url)
-                    })
-                    .render());
+        view: {
+            render() {
+                let buttons = [];
+                for (let key in this.props.pages) {
+                    const page = this.props.pages[key];
+                    buttons.push(Button.createTag()
+                        .props({
+                            label: page.title,
+                            isActive: this.props.currentPage.id === key
+                        })
+                        .methods({
+                            onclick: this.methods.navigate.bind(this, page.url)
+                        })
+                        .render());
+                }
+                return `
+                    <div class="sepcon sepcon-container flex-container">
+                        ${buttons.join('')}
+                    </div>`;
             }
-            return `
-        <div class="sepcon sepcon-container flex-container">
-            ${buttons.join('')}
-        </div>`;
         }
     }
 });
