+++ test-example-app/containers/numbers-modifier-container.js
@@ -4,10 +4,12 @@ import Counter from './../components/counter';
 export default SepCon.createComponent({
     id: 'numbers-modifier-container',
     component: {
-        render() {
-            return `<div class="sepcon sepcon-container">
-                ${Counter.createTag().render()}
-            </div>`;
+        view: {
+            render() {
+                return `<div class="sepcon sepcon-container">
+                    ${Counter.createTag().render()}
+                </div>`;
+            }
         }
     }
 });
