+++ test-example-app/containers/passing-props-container.js
@@ -1,0 +1,18 @@
+import { SepCon } from '../../src/index';
+import PropsPasser from './../components/props-passer';
+
+export default SepCon.createComponent({
+    id: 'passing-props-container',
+    component: {
+        view: {
+            render() {
+                const propsPasser = PropsPasser.createTag();
+                return `<div class="sepcon sepcon-container">
+                    <div class="presentors">
+                        ${propsPasser.render()}
+                    </div>
+                </div>`;
+            }
+        }
+    }
+});
