+++ test-example-app/elements/described-button.js
@@ -25,25 +25,27 @@ export default SepCon.createComponent({
                 }
             ]
         },
-        'events': [
-            {event: 'click', selector: '.clickable', callback: 'handleClick'}
-        ],
-        'render'() {
-            let classes = ['button', 'raised', 'clickable'];
-            if (this.props.isActive) {
-                classes.push('is-active');
-            }
-            return `<div class="sepcon sepcon-element">
+        view: {
+            events: [
+                {event: 'click', selector: '.clickable', callback: 'handleClick'}
+            ],
+            'render'() {
+                let classes = ['button', 'raised', 'clickable'];
+                if (this.props.isActive) {
+                    classes.push('is-active');
+                }
+                return `<div class="sepcon sepcon-element">
                 <div class="${classes.join(' ')}">
                     <h5>${this.props.label}</h5>
                 </div>
                 ${this.children.length ? this.children.map(el => el.outerHTML).join('') : ''}
             </div>`;
-        },
-        handleClick(e) {
-            e.preventDefault();
-            e.stopPropagation();
-            this.methods.onclick(123);
+            },
+            handleClick(e) {
+                e.preventDefault();
+                e.stopPropagation();
+                this.methods.onclick(123);
+            }
         }
     }
 });
