+++ test-example-app/layouts/layout.js
@@ -18,34 +18,36 @@ export default SepCon.createComponent({
                 }
             }
         },
-        render() {
-            if (!this.props.page) {
-                return false;
-            }
-            let header = Header.createTag();
-            let navbar = Navbar.createTag();
-            let html = '';
-            switch (this.props.page.id) {
-                case 'home':
-                    html = `<div class="sepcon sepcon-container">
+        view: {
+            render() {
+                if (!this.props.page) {
+                    return false;
+                }
+                let header = Header.createTag();
+                let navbar = Navbar.createTag();
+                let html = '';
+                switch (this.props.page.id) {
+                    case 'home':
+                        html = `<div class="sepcon sepcon-container">
                     ${NumbersModifierContainer.createTag().render()}
                 </div>`;
-                    break;
-                case 'flux':
-                    html = `<div class="sepcon sepcon-container">
+                        break;
+                    case 'flux':
+                        html = `<div class="sepcon sepcon-container">
                     ${NumbersContainer.createTag().render()}
                 </div>`;
-                    break;
-                case 'externals':
-                    html = `<div class="sepcon sepcon-container">
+                        break;
+                    case 'externals':
+                        html = `<div class="sepcon sepcon-container">
                     ${PassingPropsContainer.createTag().render()}
                 </div>`;
-                    break;
+                        break;
+                }
+                return `${header.render()}${navbar.render()}${html}`;
+            },
+            descendantChange(component) {
+                console.log('layout descendantChange', component);
             }
-            return `${header.render()}${navbar.render()}${html}`;
-        },
-        descendantChange(component) {
-            console.log('layout descendantChange', component);
         }
     }
 });
